Models now use floating point numbers for vertex coordinates.
This is because my R300 doesn't support other formats well.

+++ 3-objects/src/model.hpp
@@ -9,7 +9,7 @@ struct Model {
     void draw(const Color &tcolor, const Color &lcolor) const;
 
     GLint vertexCount;
-    const GLshort (*vertex)[3];
+    const GLfloat (*vertex)[3];
     GLsizei triCount;
     const GLubyte (*tri)[3];
     GLsizei lineCount;
